requirements.txt
streamlit>=1.21.0
langchain>=0.0.335
langchain-community>=0.0.16
langchain-core>=0.1.45
langchain-groq>=0.1.0
python-dotenv>=1.0.0
streamlit-lottie>=0.0.5
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pypdf>=3.15.0
huggingface-hub>=0.16.0
requests>=2.28.0
protobuf>=3.20.0
pydantic>=2.0.0

